Improve type knowledge (thanks to herhut-ggl).

+++ Dart.RaytracerDemo/web/missing.dart
@@ -101,7 +101,7 @@ class MersenneTwister
   
      double genrand_real1()
      {
-        return _mt.callMethod("genrand_real1");
+        return _mt.callMethod("genrand_real1") as double;
      }
 }
